Move kpoints interface to dedicated submodule
