+++ docs/api/autodoc/spglib.kpoints.rst
@@ -1,0 +1,8 @@
+spglib.kpoints module
+=====================
+
+.. automodule:: spglib.kpoints
+   :members:
+   :private-members:
+   :show-inheritance:
+   :undoc-members:
